pyproject.toml
[tool.poetry]
name = "lydia_image_similarity"
version = "0.1.0"
description = ""
authors = ["Rémi PERRIER <dev.perrier.remi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["standard"], version = "^0.112.2"}
ipykernel = "^6.29.5"
matplotlib = "^3.9.2"
pandas = "^2.2.2"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
torch = "^2.4.0"
torchvision = "^0.19.0"
tqdm = "^4.66.5"

[tool.poetry.dev-dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
